+++ lib/src/cars/presentation/widgets/car_details_carousel_items.dart
